requirements.txt
xgboost
pdfplumber
pypdf
pypdf2
huggingface_hub
ruamel.yaml
hanziconv
datrie
pyclipper
shapely
onnxruntime
opencv-python
opencv-python-headless
nltk
cachetools
pdfminer.six
six
fasttext
python-pptx
Aspose.Slides==24.2.0
tiktoken
python-docx 
pandas
readability-lxml
html_text
chardet
strenum